Move Runtime to existing collector package
